--- github-experimental-plugin/src/main/resources/checkstyle/checkstyle-suppressions.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE suppressions PUBLIC
-                "-//Puppy Crawl//DTD Suppressions 1.0//EN"
-                "http://www.puppycrawl.com/dtds/suppressions_1_0.dtd">
-
-<suppressions>
-
-    <suppress checks="JavadocStyleCheck" files="package-info.java"/>
-
-</suppressions>
