+++ github-pullrequest-plugin/pom.xml
@@ -68,7 +68,7 @@
         <dependency>
             <groupId>com.coravy.hudson.plugins.github</groupId>
             <artifactId>github</artifactId>
-            <version>1.14.2</version>
+            <version>1.19.1</version>
             <exclusions>
                 <exclusion>
                     <groupId>org.slf4j</groupId>
