+++ github-pullrequest-plugin/src/main/java/com/github/kostyasha/github/integration/branch/GitHubBranchTrigger.java
@@ -1,16 +1,18 @@
-package org.jenkinsci.plugins.github.pullrequest;
+package com.github.kostyasha.github.integration.branch;
 
 import antlr.ANTLRException;
+import com.github.kostyasha.github.integration.branch.events.GitHubBranchEvent;
+import com.github.kostyasha.github.integration.branch.trigger.JobRunnerForBranchCause;
+import com.github.kostyasha.github.integration.branch.trigger.check.NotUpdatedBranchFilter;
+import com.github.kostyasha.github.integration.branch.trigger.check.PushToCauseConverter;
+import com.github.kostyasha.github.integration.branch.trigger.check.SkipFirstRunForBranchFilter;
 import com.cloudbees.jenkins.GitHubRepositoryName;
 import com.cloudbees.jenkins.GitHubWebHook;
 import com.coravy.hudson.plugins.github.GithubProjectProperty;
-import com.google.common.base.Optional;
 import hudson.Extension;
-import hudson.Util;
 import hudson.model.Action;
 import hudson.model.Item;
 import hudson.model.Job;
-import hudson.triggers.Trigger;
 import hudson.triggers.TriggerDescriptor;
 import hudson.util.SequentialExecutionQueue;
 import jenkins.model.Jenkins;
@@ -18,18 +20,14 @@ import jenkins.model.ParameterizedJobMixIn;
 import jenkins.triggers.SCMTriggerItem;
 import net.sf.json.JSONObject;
 import org.jenkinsci.plugins.github.GitHubPlugin;
-import org.jenkinsci.plugins.github.internal.GHPluginConfigException;
-import org.jenkinsci.plugins.github.pullrequest.events.GitHubPREvent;
-import org.jenkinsci.plugins.github.pullrequest.events.GitHubPREventDescriptor;
+import org.jenkinsci.plugins.github.pullrequest.GitHubPRTriggerMode;
 import org.jenkinsci.plugins.github.pullrequest.restrictions.GitHubPRBranchRestriction;
 import org.jenkinsci.plugins.github.pullrequest.restrictions.GitHubPRUserRestriction;
-import org.jenkinsci.plugins.github.pullrequest.trigger.JobRunnerForCause;
 import org.jenkinsci.plugins.github.pullrequest.utils.LoggingTaskListenerWrapper;
 import com.github.kostyasha.github.integration.generic.GitHubTrigger;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
-import org.kohsuke.github.GHIssueState;
-import org.kohsuke.github.GHPullRequest;
+import org.kohsuke.github.GHBranch;
 import org.kohsuke.github.GHRateLimit;
 import org.kohsuke.github.GHRepository;
 import org.kohsuke.github.GitHub;
@@ -56,64 +54,33 @@ import static com.google.common.base.Charsets.UTF_8;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.base.Predicates.and;
-import static com.google.common.base.Predicates.in;
 import static com.google.common.base.Predicates.not;
 import static com.google.common.base.Predicates.notNull;
 import static java.text.DateFormat.getDateTimeInstance;
-import static org.apache.commons.lang.StringUtils.isNotBlank;
-import static org.jenkinsci.plugins.github.config.GitHubServerConfig.withHost;
+import static org.jenkinsci.plugins.github.pullrequest.GitHubPRTrigger.DescriptorImpl.githubFor;
 import static org.jenkinsci.plugins.github.pullrequest.GitHubPRTriggerMode.CRON;
 import static org.jenkinsci.plugins.github.pullrequest.GitHubPRTriggerMode.LIGHT_HOOKS;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.BranchRestrictionFilter.withBranchRestriction;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.LocalRepoUpdater.updateLocalRepo;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.NotUpdatedPRFilter.notUpdated;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.PullRequestToCauseConverter.toGitHubPRCause;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.SkipFirstRunForPRFilter.ifSkippedFirstRun;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.UserRestrictionFilter.withUserRestriction;
-import static org.jenkinsci.plugins.github.pullrequest.trigger.check.UserRestrictionPopulator.prepareUserRestrictionFilter;
 import static org.jenkinsci.plugins.github.pullrequest.utils.ObjectsUtil.isNull;
 import static org.jenkinsci.plugins.github.pullrequest.utils.ObjectsUtil.nonNull;
-import static org.jenkinsci.plugins.github.pullrequest.utils.PRHelperFunctions.extractPRNumber;
-import static org.jenkinsci.plugins.github.pullrequest.utils.PRHelperFunctions.fetchRemotePR;
 import static org.jenkinsci.plugins.github.pullrequest.webhook.WebhookInfoPredicates.withHookTriggerMode;
 import static org.jenkinsci.plugins.github.util.FluentIterableWrapper.from;
 import static org.jenkinsci.plugins.github.util.JobInfoHelpers.isBuildable;
+import static com.github.kostyasha.github.integration.branch.trigger.check.LocalRepoUpdater.updateLocalRepo;
 
 /**
- * GitHub Pull Request trigger.
- * Planned trigger modes:
- * - just incoming hooks, without persist (save PR state to local xml)
- * - hooks with persist
- * - cron run, persist
- * <p>
- * Restrictions can't have resolver, so they separate and provide security check methods:
- * - Target com.github.kostyasha.github.integration.branch restriction {@link org.jenkinsci.plugins.github.pullrequest.restrictions.GitHubPRUserRestriction}
- * - User restriction (check comments, labels, etc)
- * {@link org.jenkinsci.plugins.github.pullrequest.restrictions.GitHubPRUserRestriction}
- * (whitelist manipulations using comments is also allowed)
- * <p>
- * Event triggering is modular. Now they can be split to any events:
- * - Trigger by comment
- * - Trigger when PR opened
- * - Trigger when PR closed
- * - Trigger by label
- *
  * @author Kanstantsin Shautsou
  */
-public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(GitHubPRTrigger.class);
-    public static final String FINISH_MSG = "Finished GitHub Pull Request trigger check";
+public class GitHubBranchTrigger extends GitHubTrigger<GitHubBranchTrigger> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(GitHubBranchTrigger.class);
+    public static final String FINISH_MSG = "Finished GitHub Push trigger check";
 
     @CheckForNull
     private GitHubPRTriggerMode triggerMode = CRON;
-    @CheckForNull
-    private List<GitHubPREvent> events = new ArrayList<>();
-    /**
-     * Set PR(commit) status before build. No configurable message for it.
-     */
+
+    private List<GitHubBranchEvent> events = new ArrayList<>();
+
     private boolean preStatus = false;
     private boolean cancelQueued = false;
-    private boolean abortRunning = false;
     private boolean skipFirstRun = false;
 
     @CheckForNull
@@ -126,22 +93,19 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
     private transient GHRepository remoteRepository;
 
     @CheckForNull
-    private transient GitHubPRPollingLogAction pollingLogAction;
+    private transient GitHubBranchPollingLogAction pollingLogAction;
 
     /**
      * For groovy UI
      */
     @Restricted(value = NoExternalUse.class)
-    public GitHubPRTrigger() {
+    public GitHubBranchTrigger() {
     }
 
     @DataBoundConstructor
-    public GitHubPRTrigger(String spec,
-                           GitHubPRTriggerMode triggerMode,
-                           List<GitHubPREvent> events) throws ANTLRException {
+    public GitHubBranchTrigger(String spec, GitHubPRTriggerMode triggerMode) throws ANTLRException {
         super(spec);
         this.triggerMode = triggerMode;
-        this.events = Util.fixNull(events);
     }
 
     @DataBoundSetter
@@ -152,11 +116,6 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
     @DataBoundSetter
     public void setCancelQueued(boolean cancelQueued) {
         this.cancelQueued = cancelQueued;
-    }
-
-    @DataBoundSetter
-    public void setAbortRunning(boolean abortRunning) {
-        this.abortRunning = abortRunning;
     }
 
     @DataBoundSetter
@@ -182,20 +141,12 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
         return cancelQueued;
     }
 
-    public boolean isAbortRunning() {
-        return abortRunning;
-    }
-
     public boolean isSkipFirstRun() {
         return skipFirstRun;
     }
 
     public GitHubPRTriggerMode getTriggerMode() {
         return triggerMode;
-    }
-
-    public List<GitHubPREvent> getEvents() {
-        return events;
     }
 
     public GitHubPRUserRestriction getUserRestriction() {
@@ -208,7 +159,7 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
 
     @Override
     public void start(Job<?, ?> project, boolean newInstance) {
-        LOGGER.info("Starting GitHub Pull Request trigger for project {}", project.getName());
+        LOGGER.info("Starting GitHub Push trigger for project {}", project.getName());
         super.start(project, newInstance);
 
         if (newInstance && GitHubPlugin.configuration().isManageHooks() && withHookTriggerMode().apply(project)) {
@@ -216,7 +167,6 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
         }
     }
 
-    @Override
     public void run() {
         if (getTriggerMode() != LIGHT_HOOKS) {
             doRun(null);
@@ -227,15 +177,15 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
     public void stop() {
         //TODO clean hooks?
         if (nonNull(job)) {
-            LOGGER.info("Stopping the GitHub PR trigger for project {}", job.getFullName());
+            LOGGER.info("Stopping the Push trigger for project {}", job.getFullName());
         }
         super.stop();
     }
 
     @CheckForNull
-    public GitHubPRPollingLogAction getPollingLogAction() {
+    public GitHubBranchPollingLogAction getPollingLogAction() {
         if (isNull(pollingLogAction) && nonNull(job)) {
-            pollingLogAction = new GitHubPRPollingLogAction(job);
+            pollingLogAction = new GitHubBranchPollingLogAction(job);
         }
 
         return pollingLogAction;
@@ -258,12 +208,12 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
     /**
      * For running from external places. Goes to queue.
      */
-    public void queueRun(Job<?, ?> job, final int prNumber) {
+    public void queueRun(Job<?, ?> job, final String branch) {
         this.job = job;
         getDescriptor().queue.execute(new Runnable() {
             @Override
             public void run() {
-                doRun(prNumber);
+                doRun(branch);
             }
         });
     }
@@ -309,9 +259,9 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
     /**
      * Runs check
      *
-     * @param prNumber - PR number for check, if null - then all PRs
+     * @param branch - com.github.kostyasha.github.integration.branch for check, if null - then all PRs
      */
-    public void doRun(Integer prNumber) {
+    public void doRun(String branch) {
         if (not(isBuildable()).apply(job)) {
             LOGGER.debug("Job {} is disabled, but trigger run!", isNull(job) ? "<no job>" : job.getFullName());
             return;
@@ -322,13 +272,13 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
             return;
         }
 
-        GitHubPRRepository localRepository = job.getAction(GitHubPRRepository.class);
+        GitHubBranchRepository localRepository = job.getAction(GitHubBranchRepository.class);
         if (isNull(localRepository)) {
             LOGGER.warn("Can't get repository info, maybe project {} misconfigured?", job.getFullName());
             return;
         }
 
-        List<GitHubPRCause> causes;
+        List<GitHubBranchCause> causes;
 
         try (LoggingTaskListenerWrapper listener =
                      new LoggingTaskListenerWrapper(getPollingLogAction().getPollingLogFile(), UTF_8)) {
@@ -336,7 +286,7 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
             listener.debug("Running GitHub Pull Request trigger check for {} on {}",
                     getDateTimeInstance().format(new Date(startTime)), localRepository.getFullName());
 
-            causes = readyToBuildCauses(localRepository, listener, prNumber);
+            causes = readyToBuildCauses(localRepository, listener, branch);
 
             localRepository.saveQuetly();
 
@@ -348,46 +298,38 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
             return;
         }
 
-        from(causes).filter(new JobRunnerForCause(job, this)).toSet();
+        from(causes).filter(new JobRunnerForBranchCause(job, this)).toSet();
     }
 
     /**
-     * runs check of local (last) Repository state (list of PRs) vs current remote state
-     * - local state store only last open PRs
-     * - if last open PR <-> now closed -> should trigger only when ClosePREvent exist
-     * - last open PR <-> now changed -> trigger only
-     * - special comment in PR -> trigger
-     *
-     * @param localRepository persisted data to compare with remote state
-     * @param listener        logger to write to console and to polling log
-     * @param prNumber        pull request number to fetch only required num. Can be null
-     * @return causes which ready to be converted to job-starts. One cause per repo.
+     * @return list of causes for scheduling com.github.kostyasha.github.integration.branch builds.
      */
-    private List<GitHubPRCause> readyToBuildCauses(GitHubPRRepository localRepository,
-                                                   LoggingTaskListenerWrapper listener,
-                                                   @Nullable Integer prNumber) {
+    private List<GitHubBranchCause> readyToBuildCauses(GitHubBranchRepository localRepository,
+                                                       LoggingTaskListenerWrapper listener,
+                                                       @Nullable String branch) {
         try {
-            GitHub github = DescriptorImpl.githubFor(URI.create(localRepository.getGithubUrl()));
+            GitHub github = githubFor(URI.create(localRepository.getGithubUrl()));
             GHRateLimit rateLimitBefore = github.getRateLimit();
             listener.debug("GitHub rate limit before check: {}", rateLimitBefore);
 
-            // get local and remote list of PRs
-            //FIXME HiddenField: 'remoteRepository' hides a field? renamed to `remoteRepo`
+            // get local and remote list of branches
             GHRepository remoteRepo = getRemoteRepo();
-            Set<GHPullRequest> remotePulls = pullRequestsToCheck(prNumber, remoteRepo, localRepository);
+            Set<GHBranch> remoteBranches = branchesToCheck(branch, remoteRepo, localRepository);
 
-            Set<GHPullRequest> prepeared = from(remotePulls)
-                    .filter(notUpdated(localRepository, listener))
-                    .transform(prepareUserRestrictionFilter(localRepository, this)).toSet();
+            Set<GHBranch> prepeared = from(remoteBranches)
+                    .filter(NotUpdatedBranchFilter.notUpdated(localRepository, listener))
+//                    .transform(prepareUserRestrictionFilter(localRepository, this))
+                    .toSet();
 
-            List<GitHubPRCause> causes = from(prepeared)
+            List<GitHubBranchCause> causes = from(prepeared)
                     .filter(and(
-                            ifSkippedFirstRun(listener, skipFirstRun),
-                            withBranchRestriction(listener, branchRestriction),
-                            withUserRestriction(listener, userRestriction)
+                            SkipFirstRunForBranchFilter.ifSkippedFirstRun(listener, skipFirstRun)//,
+//                            withBranchRestriction(listener, branchRestriction),
+//                            withUserRestriction(listener, userRestriction)
                     ))
-                    .transform(toGitHubPRCause(localRepository, listener, this))
-                    .filter(notNull()).toList();
+                    .transform(PushToCauseConverter.toGitHubPushCause(localRepository, listener, this))
+                    .filter(notNull())
+                    .toList();
 
             LOGGER.trace("Causes count for {}: {}", localRepository.getFullName(), causes.size());
             from(prepeared).transform(updateLocalRepo(localRepository)).toSet();
@@ -397,13 +339,18 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
             GHRateLimit rateLimitAfter = github.getRateLimit();
             int consumed = rateLimitBefore.remaining - rateLimitAfter.remaining;
             LOGGER.info("GitHub rate limit after check {}: {}, consumed: {}, checked PRs: {}",
-                    localRepository.getFullName(), rateLimitAfter, consumed, remotePulls.size());
+                    localRepository.getFullName(), rateLimitAfter, consumed, remoteBranches.size());
 
             return causes;
         } catch (IOException e) {
             listener.error("Can't get build causes, because: '{}'", e.getMessage());
             return Collections.emptyList();
         }
+    }
+
+    private Set<GHBranch> branchesToCheck(String branch, GHRepository remoteRepo, GitHubBranchRepository localRepository)
+            throws IOException {
+        return (Set<GHBranch>) remoteRepo.getBranches().values();
     }
 
     private void saveIfSkipFirstRun() {
@@ -418,26 +365,6 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
         return mode != LIGHT_HOOKS;
     }
 
-    /**
-     * @return remote pull requests for future analysing.
-     */
-    private static Set<GHPullRequest> pullRequestsToCheck(@Nullable Integer prNumber,
-                                                          GHRepository remoteRepo,
-                                                          GitHubPRRepository localRepo) throws IOException {
-        if (prNumber != null) {
-            return Collections.singleton(remoteRepo.getPullRequest(prNumber));
-        } else {
-            List<GHPullRequest> remotePulls = remoteRepo.getPullRequests(GHIssueState.OPEN);
-
-            Set<Integer> remotePRNums = from(remotePulls).transform(extractPRNumber()).toSet();
-
-            return from(localRepo.getPulls().keySet())
-                    // add PRs that was closed on remote
-                    .filter(not(in(remotePRNums)))
-                    .transform(fetchRemotePR(remoteRepo)).filter(notNull()).append(remotePulls).toSet();
-        }
-    }
-
     @CheckForNull
     public Job<?, ?> getJob() {
         return job;
@@ -447,8 +374,6 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
     public static class DescriptorImpl extends TriggerDescriptor {
         private final transient SequentialExecutionQueue queue =
                 new SequentialExecutionQueue(Jenkins.MasterComputer.threadPoolForRemoting);
-
-        private String publishedURL;
 
         public DescriptorImpl() {
             load();
@@ -462,7 +387,7 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
 
         @Override
         public String getDisplayName() {
-            return "Build GitHub pull requests";
+            return "Experimental: Build GitHub branches";
         }
 
         @Override
@@ -473,43 +398,8 @@ public class GitHubPRTrigger extends GitHubTrigger<GitHubPRTrigger> {
             return super.configure(req, formData);
         }
 
-        public String getPublishedURL() {
-            return publishedURL;
-        }
-
-        public void setPublishedURL(String publishedURL) {
-            this.publishedURL = publishedURL;
-        }
-
-        public String getJenkinsURL() {
-            String url = getPublishedURL();
-            if (isNotBlank(url)) {
-                if (!url.endsWith("/")) {
-                    url += "/";
-                }
-                return url;
-            }
-            return GitHubWebHook.getJenkinsInstance().getRootUrl();
-        }
-
-        // list all available descriptors for choosing in job configuration
-        public List<GitHubPREventDescriptor> getEventDescriptors() {
-            return GitHubPREventDescriptor.all();
-        }
-
-        @Nonnull
-        public static GitHub githubFor(URI uri) {
-            Optional<GitHub> client = from(GitHubPlugin.configuration()
-                    .findGithubConfig(withHost(uri.getHost()))).first();
-            if (client.isPresent()) {
-                return client.get();
-            } else {
-                throw new GHPluginConfigException("Can't find appropriate client for github repo <%s>", uri);
-            }
-        }
-
         public static DescriptorImpl get() {
-            return Trigger.all().get(DescriptorImpl.class);
+            return all().get(DescriptorImpl.class);
         }
     }
 }
