+++ github-pullrequest-plugin/src/main/java/com/github/kostyasha/github/integration/branch/trigger/check/SkipFirstRunForBranchFilter.java
@@ -1,0 +1,32 @@
+package com.github.kostyasha.github.integration.branch.trigger.check;
+
+import com.google.common.base.Predicate;
+import org.jenkinsci.plugins.github.pullrequest.utils.LoggingTaskListenerWrapper;
+import org.kohsuke.github.GHBranch;
+
+/**
+ * @author Kanstantsin Shautsou
+ */
+public class SkipFirstRunForBranchFilter implements Predicate<GHBranch> {
+    private final LoggingTaskListenerWrapper logger;
+    private final boolean skipFirstRun;
+
+    private SkipFirstRunForBranchFilter(LoggingTaskListenerWrapper logger, boolean skipFirstRun) {
+        this.logger = logger;
+        this.skipFirstRun = skipFirstRun;
+    }
+
+    public static Predicate<GHBranch> ifSkippedFirstRun(LoggingTaskListenerWrapper logger, boolean skipFirstRun) {
+        return new SkipFirstRunForBranchFilter(logger, skipFirstRun);
+    }
+
+    @Override
+    public boolean apply(GHBranch remoteBranch) {
+        if (skipFirstRun) {
+            logger.info("Skipping first run for com.github.kostyasha.github.integration.branch #{}", remoteBranch.getName());
+            return false;
+        }
+
+        return true;
+    }
+}
