+++ github-pullrequest-plugin/src/main/java/org/jenkinsci/plugins/github/pullrequest/GitHubPRCause.java
@@ -1,8 +1,6 @@
 package org.jenkinsci.plugins.github.pullrequest;
 
-import hudson.model.Cause;
-import hudson.model.Run;
-import org.apache.commons.io.FileUtils;
+import com.github.kostyasha.github.integration.generic.GitHubCause;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
@@ -15,7 +13,6 @@ import org.slf4j.LoggerFactory;
 
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
-import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.util.Collections;
@@ -24,7 +21,7 @@ import java.util.Set;
 import static org.jenkinsci.plugins.github.pullrequest.utils.ObjectsUtil.isNull;
 import static org.jenkinsci.plugins.github.pullrequest.utils.ObjectsUtil.nonNull;
 
-public class GitHubPRCause extends Cause {
+public class GitHubPRCause extends GitHubCause {
     private static final Logger LOGGER = LoggerFactory.getLogger(GitHubPRCause.class);
 
     private String headSha;
@@ -35,7 +32,6 @@ public class GitHubPRCause extends Cause {
     private String prAuthorEmail;
     @CheckForNull
     private String title;
-    private URL htmlUrl;
     private String sourceRepoOwner;
     private String triggerSenderName = "";
     private String triggerSenderEmail = "";
@@ -48,9 +44,7 @@ public class GitHubPRCause extends Cause {
     private String commitAuthorName;
     private String commitAuthorEmail;
 
-    private boolean skip;
     private String condRef;
-    private String pollingLog;
     private String state;
 
     public GitHubPRCause() {
@@ -81,40 +75,40 @@ public class GitHubPRCause extends Cause {
                          GHUser triggerSender, boolean skip, String reason,
                          String commitAuthorName, String commitAuthorEmail,
                          String state) {
-    //CHECKSTYLE:ON
-        this.headSha = headSha;
-        this.number = number;
-        this.mergeable = mergeable;
-        this.targetBranch = targetBranch;
-        this.sourceBranch = sourceBranch;
-        this.prAuthorEmail = prAuthorEmail;
-        this.title = title;
-        this.htmlUrl = htmlUrl;
-        this.sourceRepoOwner = sourceRepoOwner;
-        this.labels = labels;
-        this.skip = skip;
-        this.reason = reason;
-        this.commitAuthorName = commitAuthorName;
-        this.commitAuthorEmail = commitAuthorEmail;
+            //CHECKSTYLE:ON
+            this.headSha = headSha;
+            this.number = number;
+            this.mergeable = mergeable;
+            this.targetBranch = targetBranch;
+            this.sourceBranch = sourceBranch;
+            this.prAuthorEmail = prAuthorEmail;
+            this.title = title;
+            this.htmlUrl = htmlUrl;
+            this.sourceRepoOwner = sourceRepoOwner;
+            this.labels = labels;
+            this.skip = skip;
+            this.reason = reason;
+            this.commitAuthorName = commitAuthorName;
+            this.commitAuthorEmail = commitAuthorEmail;
 
-        if (nonNull(triggerSender)) {
-            try {
-                this.triggerSenderName = triggerSender.getName();
-            } catch (IOException e) {
-                LOGGER.error("Can't get trigger sender name from remote PR");
+            if (nonNull(triggerSender)) {
+                try {
+                    this.triggerSenderName = triggerSender.getName();
+                } catch (IOException e) {
+                    LOGGER.error("Can't get trigger sender name from remote PR");
+                }
+
+                try {
+                    this.triggerSenderEmail = triggerSender.getEmail();
+                } catch (IOException e) {
+                    LOGGER.error("Can't get trigger sender email from remote PR");
+                }
             }
 
-            try {
-                this.triggerSenderEmail = triggerSender.getEmail();
-            } catch (IOException e) {
-                LOGGER.error("Can't get trigger sender email from remote PR");
-            }
+            this.condRef = mergeable ? "merge" : "head";
+
+            this.state = state;
         }
-
-        this.condRef = mergeable ? "merge" : "head";
-
-        this.state = state;
-    }
 
 
     public static GitHubPRCause newGitHubPRCause() {
@@ -178,14 +172,6 @@ public class GitHubPRCause extends Cause {
     }
 
     /**
-     * @see #htmlUrl
-     */
-    public GitHubPRCause withHtmlUrl(URL htmlUrl) {
-        this.htmlUrl = htmlUrl;
-        return this;
-    }
-
-    /**
      * @see #sourceRepoOwner
      */
     public GitHubPRCause withSourceRepoOwner(String sourceRepoOwner) {
@@ -242,40 +228,11 @@ public class GitHubPRCause extends Cause {
     }
 
     /**
-     * @see #skip
-     */
-    public GitHubPRCause withSkip(boolean skip) {
-        this.skip = skip;
-        return this;
-    }
-
-    /**
      * @see #condRef
      */
     public GitHubPRCause withCondRef(String condRef) {
         this.condRef = condRef;
         return this;
-    }
-
-    /**
-     * @see #pollingLog
-     */
-    public GitHubPRCause withPollingLog(String pollingLog) {
-        this.pollingLog = pollingLog;
-        return this;
-    }
-
-    @Override
-    public void onAddedTo(@Nonnull Run run) {
-        // move polling log from cause to action
-        try {
-            GitHubPRPollingLogAction action = new GitHubPRPollingLogAction(run);
-            FileUtils.writeStringToFile(action.getPollingLogFile(), pollingLog);
-            run.replaceAction(action);
-        } catch (IOException e) {
-            LOGGER.warn("Failed to persist the polling log", e);
-        }
-        pollingLog = null;
     }
 
     @Override
@@ -307,11 +264,6 @@ public class GitHubPRCause extends Cause {
         return prAuthorEmail;
     }
 
-    // for printing PR url on left builds panel (build description)
-    public URL getHtmlUrl() {
-        return htmlUrl;
-    }
-
     public String getSourceRepoOwner() {
         return sourceRepoOwner;
     }
@@ -327,10 +279,6 @@ public class GitHubPRCause extends Cause {
 
     public String getTriggerSenderEmail() {
         return triggerSenderEmail;
-    }
-
-    public boolean isSkip() {
-        return skip;
     }
 
     public String getReason() {
@@ -371,14 +319,6 @@ public class GitHubPRCause extends Cause {
     @Nonnull
     public String getCondRef() {
         return condRef;
-    }
-
-    public void setPollingLog(String pollingLog) {
-        this.pollingLog = pollingLog;
-    }
-
-    public void setPollingLog(File logFile) throws IOException {
-        this.pollingLog = FileUtils.readFileToString(logFile);
     }
 
     @Override
