--- github-pullrequest-plugin/src/main/java/org/jenkinsci/plugins/github_integration/generic/GitHubAbstractRepository.java
@@ -1,10 +1,0 @@
-package org.jenkinsci.plugins.github_integration.generic;
-
-import hudson.model.Action;
-import hudson.model.Saveable;
-
-/**
- * @author Kanstantsin Shautsou
- */
-public abstract class GitHubAbstractRepository implements Action, Saveable {
-}
