+++ github-pullrequest-plugin/src/main/resources/org/jenkinsci/plugins/github/pullrequest/GitHubPRTrigger/help-branchRestriction.html
@@ -1,3 +1,3 @@
 <div>
-    Limit job run to certain branch matches.
+    Limit job run to certain com.github.kostyasha.github.integration.branch matches.
 </div>
